requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pandas==2.2.3
datasets==3.1.0
requests==2.32.5
streamlit==1.40.0
streamlit-extras==0.4.7
